requirements.txt
djangorestframework
django-ckeditor-5
gunicorn
Django
djangorestframework-simplejwt
django-filter
django-jazzmin
Pillow
hashids
django-cors-headers
